Estimativas prontas, só falta 2 médias

+++ DemandaFacil/src/models/estimativa/Estimativas.java
@@ -3,14 +3,14 @@ package models.estimativa;
 public class Estimativas {
    private String data;
    private int Produto_idProduto;
-   private double valorCalculado;
+   private int valorCalculado;
 
    public Estimativas(){
        
        
    } 
    
-   public Estimativas(String data, int Produto_idProduto, double valorCalculado) {
+   public Estimativas(String data, int Produto_idProduto, int valorCalculado) {
         this.data = data;
         this.Produto_idProduto = Produto_idProduto;
         this.valorCalculado = valorCalculado;
@@ -32,11 +32,11 @@ public class Estimativas {
         this.Produto_idProduto = Produto_idProduto;
     }
 
-    public double getValorCalculado() {
+    public int getValorCalculado() {
         return valorCalculado;
     }
 
-    public void setValorCalculado(double valorCalculado) {
+    public void setValorCalculado(int valorCalculado) {
         this.valorCalculado = valorCalculado;
     }
